build: update ts defs and docs parser
